+++ src/Exceptions/GoogleAuthenticationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Statikbe\GoogleAuthenticate\Exceptions;
+
+class GoogleAuthenticationException extends \Exception
+{
+}